+++ examples/data-sources/sse_tenant_controls_profile/data-source.tf
@@ -1,0 +1,7 @@
+data "sse_tenant_controls_profile" "example" {
+  name = "Global Tenant Controls"
+}
+
+output "tenant_controls_profile_id" {
+  value = data.sse_tenant_controls_profile.example.id
+}