+++ examples/data-sources/sse_tenant_controls_profiles/data-source.tf
@@ -1,0 +1,5 @@
+data "sse_tenant_controls_profiles" "all" {}
+
+output "all_tenant_controls_profiles" {
+  value = data.sse_tenant_controls_profiles.all.profiles
+}